requirements/base.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
channels==4.0.0
channels-redis==4.2.0
uWSGI>=2.0.19.1,<2.1
celery==5.3.6
redis==5.0.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
kavenegar==1.1.2
django-ratelimit==4.1.0
Pillow==10.2.0